Move class 'org.hibernate.tool.hbm2x.QueryExporter' to 'org.hibernate.tool.internal.export.query.QueryExporter'
